requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
Deprecated==1.2.15
distro==1.9.0
Flask==3.1.0
Flask-BasicAuth==0.2.0
Flask-Limiter==3.10.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.59.7
ordered-set==4.1.0
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
gunicorn==21.2.0